# target/classes/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Todo List API
  description: |
    RESTful API for managing a personal todo list.
    Supports full CRUD operations on todos with priority, due dates, and tag support.
  version: 1.0.0
  contact:
    name: Cypcode Studios
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server

tags:
  - name: todos
    description: Todo item operations
  - name: tags
    description: Tag management

paths:
  /todos:
    get:
      tags: [todos]
      summary: List all todos
      operationId: listTodos
      parameters:
        - name: status
          in: query
          description: Filter by completion status
          schema:
            type: string
            enum: [all, active, completed]
            default: all
        - name: priority
          in: query
          description: Filter by priority level
          schema:
            $ref: '#/components/schemas/Priority'
        - name: tag
          in: query
          description: Filter by tag name
          schema:
            type: string
        - name: search
          in: query
          description: Full-text search on title and description
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [createdAt, dueDate, priority, title]
            default: createdAt
        - name: sortDir
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Paginated list of todos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoPage'

    post:
      tags: [todos]
      summary: Create a new todo
      operationId: createTodo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTodoRequest'
            example:
              title: Buy groceries
              description: Milk, eggs, bread
              priority: MEDIUM
              dueDate: '2026-05-30'
              tags: [shopping, errands]
      responses:
        '201':
          description: Todo created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          $ref: '#/components/responses/BadRequest'

  /todos/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64

    get:
      tags: [todos]
      summary: Get a todo by ID
      operationId: getTodoById
      responses:
        '200':
          description: Todo found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [todos]
      summary: Replace a todo entirely
      operationId: updateTodo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTodoRequest'
      responses:
        '200':
          description: Todo updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [todos]
      summary: Partially update a todo
      operationId: patchTodo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchTodoRequest'
      responses:
        '200':
          description: Todo patched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [todos]
      summary: Delete a todo
      operationId: deleteTodo
      responses:
        '204':
          description: Todo deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'

  /todos/{id}/complete:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    patch:
      tags: [todos]
      summary: Mark a todo as completed
      operationId: completeTodo
      responses:
        '200':
          description: Todo marked as completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  /todos/{id}/reopen:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    patch:
      tags: [todos]
      summary: Reopen a completed todo
      operationId: reopenTodo
      responses:
        '200':
          description: Todo reopened
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '404':
          $ref: '#/components/responses/NotFound'

  /todos/bulk-delete:
    post:
      tags: [todos]
      summary: Delete multiple todos at once
      operationId: bulkDeleteTodos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              properties:
                ids:
                  type: array
                  items:
                    type: integer
                    format: int64
                  minItems: 1
      responses:
        '204':
          description: Todos deleted

  /todos/stats:
    get:
      tags: [todos]
      summary: Get aggregate statistics
      operationId: getTodoStats
      responses:
        '200':
          description: Statistics summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoStats'

  /tags:
    get:
      tags: [tags]
      summary: List all tags in use
      operationId: listTags
      responses:
        '200':
          description: Array of tag names
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

components:
  schemas:
    Priority:
      type: string
      enum: [LOW, MEDIUM, HIGH, URGENT]

    Todo:
      type: object
      required: [id, title, completed, priority, createdAt, updatedAt]
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 2000
          nullable: true
        completed:
          type: boolean
          default: false
        priority:
          $ref: '#/components/schemas/Priority'
        dueDate:
          type: string
          format: date
          nullable: true
        tags:
          type: array
          items:
            type: string
          uniqueItems: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
        completedAt:
          type: string
          format: date-time
          nullable: true
          readOnly: true

    CreateTodoRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 2000
        priority:
          $ref: '#/components/schemas/Priority'
          default: MEDIUM
        dueDate:
          type: string
          format: date
          nullable: true
        tags:
          type: array
          items:
            type: string
          uniqueItems: true

    UpdateTodoRequest:
      type: object
      required: [title, priority]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 2000
          nullable: true
        priority:
          $ref: '#/components/schemas/Priority'
        dueDate:
          type: string
          format: date
          nullable: true
        tags:
          type: array
          items:
            type: string
          uniqueItems: true

    PatchTodoRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 2000
          nullable: true
        priority:
          $ref: '#/components/schemas/Priority'
        dueDate:
          type: string
          format: date
          nullable: true
        tags:
          type: array
          items:
            type: string
          uniqueItems: true

    TodoPage:
      type: object
      required: [content, totalElements, totalPages, page, size]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Todo'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        page:
          type: integer
        size:
          type: integer
        first:
          type: boolean
        last:
          type: boolean

    TodoStats:
      type: object
      properties:
        total:
          type: integer
        completed:
          type: integer
        active:
          type: integer
        overdue:
          type: integer
        byPriority:
          type: object
          additionalProperties:
            type: integer

    ApiError:
      type: object
      required: [status, message, timestamp]
      properties:
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        fieldErrors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string

  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: State conflict (e.g. already completed)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'