# target/generated-sources/openapi/src/main/java/com/cypcode/todo/api/TodosApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.4.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.cypcode.todo.api;

import com.cypcode.todo.model.ApiError;
import com.cypcode.todo.model.BulkDeleteTodosRequest;
import com.cypcode.todo.model.CreateTodoRequest;
import com.cypcode.todo.model.PatchTodoRequest;
import com.cypcode.todo.model.Priority;
import com.cypcode.todo.model.Todo;
import com.cypcode.todo.model.TodoPage;
import com.cypcode.todo.model.TodoStats;
import com.cypcode.todo.model.UpdateTodoRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-06-06T11:21:33.992061+02:00[Africa/Johannesburg]", comments = "Generator version: 7.4.0")
@Validated
@Tag(name = "todos", description = "Todo item operations")
public interface TodosApi {

    /**
     * POST /todos/bulk-delete : Delete multiple todos at once
     *
     * @param bulkDeleteTodosRequest  (required)
     * @return Todos deleted (status code 204)
     */
    @Operation(
        operationId = "bulkDeleteTodos",
        summary = "Delete multiple todos at once",
        tags = { "todos" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Todos deleted")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/todos/bulk-delete",
        consumes = { "application/json" }
    )
    
    ResponseEntity<Void> bulkDeleteTodos(
        @Parameter(name = "BulkDeleteTodosRequest", description = "", required = true) @Valid @RequestBody BulkDeleteTodosRequest bulkDeleteTodosRequest
    );


    /**
     * PATCH /todos/{id}/complete : Mark a todo as completed
     *
     * @param id  (required)
     * @return Todo marked as completed (status code 200)
     *         or Resource not found (status code 404)
     *         or State conflict (e.g. already completed) (status code 409)
     */
    @Operation(
        operationId = "completeTodo",
        summary = "Mark a todo as completed",
        tags = { "todos" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Todo marked as completed", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Todo.class))
            }),
            @ApiResponse(responseCode = "404", description = "Resource not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiError.class))
            }),
            @ApiResponse(responseCode = "409", description = "State conflict (e.g. already completed)", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiError.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PATCH,
        value = "/todos/{id}/complete",
        produces = { "application/json" }
    )
    
    ResponseEntity<Todo> completeTodo(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    );


    /**
     * POST /todos : Create a new todo
     *
     * @param createTodoRequest  (required)
     * @return Todo created successfully (status code 201)
     *         or Validation error (status code 400)
     */
    @Operation(
        operationId = "createTodo",
        summary = "Create a new todo",
        tags = { "todos" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Todo created successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Todo.class))
            }),
            @ApiResponse(responseCode = "400", description = "Validation error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiError.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/todos",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<Todo> createTodo(
        @Parameter(name = "CreateTodoRequest", description = "", required = true) @Valid @RequestBody CreateTodoRequest createTodoRequest
    );


    /**
     * DELETE /todos/{id} : Delete a todo
     *
     * @param id  (required)
     * @return Todo deleted successfully (status code 204)
     *         or Resource not found (status code 404)
     */
    @Operation(
        operationId = "deleteTodo",
        summary = "Delete a todo",
        tags = { "todos" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Todo deleted successfully"),
            @ApiResponse(responseCode = "404", description = "Resource not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiError.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/todos/{id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<Void> deleteTodo(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    );


    /**
     * GET /todos/{id} : Get a todo by ID
     *
     * @param id  (required)
     * @return Todo found (status code 200)
     *         or Resource not found (status code 404)
     */
    @Operation(
        operationId = "getTodoById",
        summary = "Get a todo by ID",
        tags = { "todos" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Todo found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Todo.class))
            }),
            @ApiResponse(responseCode = "404", description = "Resource not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiError.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/todos/{id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<Todo> getTodoById(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    );


    /**
     * GET /todos/stats : Get aggregate statistics
     *
     * @return Statistics summary (status code 200)
     */
    @Operation(
        operationId = "getTodoStats",
        summary = "Get aggregate statistics",
        tags = { "todos" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Statistics summary", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TodoStats.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/todos/stats",
        produces = { "application/json" }
    )
    
    ResponseEntity<TodoStats> getTodoStats(
        
    );


    /**
     * GET /todos : List all todos
     *
     * @param status Filter by completion status (optional, default to all)
     * @param priority Filter by priority level (optional)
     * @param tag Filter by tag name (optional)
     * @param search Full-text search on title and description (optional)
     * @param sortBy  (optional, default to createdAt)
     * @param sortDir  (optional, default to desc)
     * @param page  (optional, default to 0)
     * @param size  (optional, default to 20)
     * @return Paginated list of todos (status code 200)
     */
    @Operation(
        operationId = "listTodos",
        summary = "List all todos",
        tags = { "todos" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Paginated list of todos", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TodoPage.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/todos",
        produces = { "application/json" }
    )
    
    ResponseEntity<TodoPage> listTodos(
        @Parameter(name = "status", description = "Filter by completion status", in = ParameterIn.QUERY) @Valid @RequestParam(value = "status", required = false, defaultValue = "all") String status,
        @Parameter(name = "priority", description = "Filter by priority level", in = ParameterIn.QUERY) @Valid @RequestParam(value = "priority", required = false) Priority priority,
        @Parameter(name = "tag", description = "Filter by tag name", in = ParameterIn.QUERY) @Valid @RequestParam(value = "tag", required = false) String tag,
        @Parameter(name = "search", description = "Full-text search on title and description", in = ParameterIn.QUERY) @Valid @RequestParam(value = "search", required = false) String search,
        @Parameter(name = "sortBy", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "sortBy", required = false, defaultValue = "createdAt") String sortBy,
        @Parameter(name = "sortDir", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "sortDir", required = false, defaultValue = "desc") String sortDir,
        @Min(0) @Parameter(name = "page", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "0") Integer page,
        @Min(1) @Max(100) @Parameter(name = "size", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "20") Integer size
    );


    /**
     * PATCH /todos/{id} : Partially update a todo
     *
     * @param id  (required)
     * @param patchTodoRequest  (required)
     * @return Todo patched (status code 200)
     *         or Validation error (status code 400)
     *         or Resource not found (status code 404)
     */
    @Operation(
        operationId = "patchTodo",
        summary = "Partially update a todo",
        tags = { "todos" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Todo patched", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Todo.class))
            }),
            @ApiResponse(responseCode = "400", description = "Validation error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiError.class))
            }),
            @ApiResponse(responseCode = "404", description = "Resource not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiError.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PATCH,
        value = "/todos/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<Todo> patchTodo(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id,
        @Parameter(name = "PatchTodoRequest", description = "", required = true) @Valid @RequestBody PatchTodoRequest patchTodoRequest
    );


    /**
     * PATCH /todos/{id}/reopen : Reopen a completed todo
     *
     * @param id  (required)
     * @return Todo reopened (status code 200)
     *         or Resource not found (status code 404)
     */
    @Operation(
        operationId = "reopenTodo",
        summary = "Reopen a completed todo",
        tags = { "todos" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Todo reopened", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Todo.class))
            }),
            @ApiResponse(responseCode = "404", description = "Resource not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiError.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PATCH,
        value = "/todos/{id}/reopen",
        produces = { "application/json" }
    )
    
    ResponseEntity<Todo> reopenTodo(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    );


    /**
     * PUT /todos/{id} : Replace a todo entirely
     *
     * @param id  (required)
     * @param updateTodoRequest  (required)
     * @return Todo updated (status code 200)
     *         or Validation error (status code 400)
     *         or Resource not found (status code 404)
     */
    @Operation(
        operationId = "updateTodo",
        summary = "Replace a todo entirely",
        tags = { "todos" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Todo updated", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Todo.class))
            }),
            @ApiResponse(responseCode = "400", description = "Validation error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiError.class))
            }),
            @ApiResponse(responseCode = "404", description = "Resource not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ApiError.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/todos/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<Todo> updateTodo(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id,
        @Parameter(name = "UpdateTodoRequest", description = "", required = true) @Valid @RequestBody UpdateTodoRequest updateTodoRequest
    );

}